+++ changelog.md
@@ -3,6 +3,7 @@
 ## Bugfix
 
 - round light color of all rules in ``habapp_rules.actors.light_hcl`` to integer values to avoid strange formating in OpenHAB
+- added config parameter ``leaving_only_if_on`` to ``habapp_rules.actors.config.light.LightParameter`` to disable unexpected leaving light, if light was not on and leaving started
 
 # Version 6.0.0 - 27.06.2024